+++ .eslintrc
@@ -21,7 +21,7 @@
   "rules": {
     "indent": [
       "warn",
-      4
+      2
     ],
     "quotes": [
       "error",
